feat: move server management to per-server settings modal (#96)
- Remove global server settings cog from sidebar bottom
- Add hover-revealed settings cog to each server icon
- Create ServerSettingsModal component using existing Modal
- Clicking cog on inactive server switches to it before opening modal
- Modal loads server data on open and is permission-aware
